unsigned long long ft_size

+++ ft_size_nb.c
@@ -6,18 +6,18 @@
 /*   By: alafranc <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/14 11:16:03 by alafranc          #+#    #+#             */
-/*   Updated: 2020/12/14 11:16:35 by alafranc         ###   ########lyon.fr   */
+/*   Updated: 2020/12/14 11:25:39 by alafranc         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_size_nb(long nb, int base)
+int	ft_size_nb(unsigned long long nb, int base)
 {
 	int count;
 
 	count = 1;
-	while (nb >= base)
+	while (nb >= (unsigned long long)base)
 	{
 		nb /= base;
 		count++;
